type tidying
